requirements.txt
# Core HTTP and API libraries
requests>=2.31.0
urllib3>=1.21.1

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b

# AI and Machine Learning
groq>=0.4.0

# Authentication and Security
PyJWT>=2.8.0
cryptography>=41.0.0

# Data visualization and charting
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web automation and screenshot generation
playwright>=1.40.0

# Scheduling and time management
schedule>=1.2.0

# Date and time handling
python-dateutil>=2.8.2

# JSON handling
jsonschema>=4.17.0

# Environment and configuration
python-dotenv>=1.0.0

# Optional: For enhanced data analysis
scipy>=1.10.0
scikit-learn>=1.3.0

# Optional: For advanced charting
mplfinance>=0.12.9b

# Optional: For web scraping (if needed)
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Optional: For database operations (if needed)
sqlalchemy>=2.0.0

# Optional: For async operations (if needed)
aiohttp>=3.8.0
asyncio>=3.4.3

# Optional: For logging and monitoring
loguru>=0.7.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 